Feature/business logic

+++ src/main/java/io/hhplus/concert/reservation/presentation/request/PaymentRequest.java
@@ -1,10 +1,12 @@
 package io.hhplus.concert.reservation.presentation.request;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class PaymentRequest {
     private String reservationId;
-    private int amount;
+    private BigDecimal amount;
     private String token;
 }
